move recorder, logger to other mods
